Strengthen foreign key constraints now that I can test on PostgreSQL

+++ chordial/models/mixins.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, String
 
 from chordial.utils.uid import encode, generate_id
 
 def id_mixin(length: int):
   class IdMixin:
-    id = Column(Integer, primary_key=True, default=lambda: generate_id(length))
+    id = Column(BigInteger, primary_key=True, default=lambda: generate_id(length))
 
     def __repr__(self):
       return f"<{type(self).__name__} {self.repr_label} ({self.uid})>"
